moved base class to core
