+++ libs/langchain/langchain/agents/output_parsers/__init__.py
@@ -9,9 +9,6 @@ This contains a `return_values` dictionary. This usually contains a
 single `output` key, but can be extended to contain more.
 This also contains a `log` variable (which contains a log of the agent's thinking).
 """
-from langchain.agents.output_parsers.functions_utils import (
-    BaseFunctionsAgentOutputParser,
-)
 from langchain.agents.output_parsers.json import JSONAgentOutputParser
 from langchain.agents.output_parsers.openai_functions import (
     OpenAIFunctionsAgentOutputParser,
@@ -26,7 +23,6 @@ from langchain.agents.output_parsers.self_ask import SelfAskOutputParser
 from langchain.agents.output_parsers.xml import XMLAgentOutputParser
 
 __all__ = [
-    "BaseFunctionsAgentOutputParser",
     "ReActSingleInputOutputParser",
     "SelfAskOutputParser",
     "ReActJsonSingleInputOutputParser",
